+++ vulnerabilities/importers/gentoo.py
@@ -10,73 +10,72 @@
 
 import re
 import xml.etree.ElementTree as ET
-from typing import Set
+from pathlib import Path
+from typing import Iterable
 
 from packageurl import PackageURL
+from univers.version_constraint import VersionConstraint
+from univers.version_range import EbuildVersionRange
+from univers.versions import GentooVersion
 
 from vulnerabilities.importer import AdvisoryData
-from vulnerabilities.importer import GitImporter
+from vulnerabilities.importer import AffectedPackage
+from vulnerabilities.importer import Importer
 from vulnerabilities.importer import Reference
-from vulnerabilities.utils import nearest_patched_package
 
 
-class GentooImporter(GitImporter):
-    def __enter__(self):
-        super(GentooImporter, self).__enter__()
+class GentooImporter(Importer):
+    repo_url = "git+https://anongit.gentoo.org/git/data/glsa.git"
+    spdx_license_expression = "CC-BY-SA-4.0"
+    # the license notice is at this url https://anongit.gentoo.org/ says:
+    # The contents of this document, unless otherwise expressly stated, are licensed
+    # under the [CC-BY-SA-4.0](https://creativecommons.org/licenses/by-sa/4.0/) license.
+    license_url = "https://creativecommons.org/licenses/by-sa/4.0/"
 
-        if not getattr(self, "_added_files", None):
-            self._added_files, self._updated_files = self.file_changes(
-                recursive=True, file_ext="xml"
-            )
-
-    def updated_advisories(self) -> Set[AdvisoryData]:
-        files = self._updated_files.union(self._added_files)
-        advisories = []
-        for f in files:
-            processed_data = self.process_file(f)
-            advisories.extend(processed_data)
-        return self.batch_advisories(advisories)
+    def advisory_data(self) -> Iterable[AdvisoryData]:
+        try:
+            self.clone(repo_url=self.repo_url)
+            base_path = Path(self.vcs_response.dest_dir)
+            for file_path in base_path.glob("**/*.xml"):
+                yield from self.process_file(file_path)
+        finally:
+            if self.vcs_response:
+                self.vcs_response.delete()
 
     def process_file(self, file):
-        xml_data = {}
+        cves = []
+        summary = ""
+        vuln_references = []
         xml_root = ET.parse(file).getroot()
-        glsa = "GLSA-" + xml_root.attrib["id"]
-        vuln_reference = [
-            Reference(
-                reference_id=glsa,
-                url="https://security.gentoo.org/glsa/{}".format(xml_root.attrib["id"]),
-            )
-        ]
+        id = xml_root.attrib.get("id")
+        if id:
+            glsa = "GLSA-" + id
+            vuln_references = [
+                Reference(
+                    reference_id=glsa,
+                    url=f"https://security.gentoo.org/glsa/{id}",
+                )
+            ]
 
         for child in xml_root:
             if child.tag == "references":
-                xml_data["cves"] = self.cves_from_reference(child)
+                cves = self.cves_from_reference(child)
 
             if child.tag == "synopsis":
-                xml_data["description"] = child.text
+                summary = child.text
 
             if child.tag == "affected":
-                (
-                    xml_data["affected_purls"],
-                    xml_data["unaffected_purls"],
-                ) = self.affected_and_safe_purls(child)
-                xml_data["unaffected_purls"] = list(xml_data["unaffected_purls"])
-                xml_data["affected_purls"] = list(xml_data["affected_purls"])
+                affected_packages = list(self.affected_and_safe_purls(child))
 
-        advisory_list = []
         # It is very inefficient, to create new Advisory for each CVE
         # this way, but there seems no alternative.
-        for cve in xml_data["cves"]:
-            advisory = AdvisoryData(
-                vulnerability_id=cve,
-                summary=xml_data["description"],
-                affected_packages=nearest_patched_package(
-                    xml_data["affected_purls"], xml_data["unaffected_purls"]
-                ),
-                references=vuln_reference,
+        for cve in cves:
+            yield AdvisoryData(
+                aliases=[cve],
+                summary=summary,
+                references=vuln_references,
+                affected_packages=affected_packages,
             )
-            advisory_list.append(advisory)
-        return advisory_list
 
     @staticmethod
     def cves_from_reference(reference):
@@ -91,40 +90,63 @@ class GentooImporter(GitImporter):
 
     @staticmethod
     def affected_and_safe_purls(affected_elem):
-        safe_purls = set()
-        affected_purls = set()
-        skip_versions = {"1.3*", "7.3*", "7.4*"}
+        constraints = []
         for pkg in affected_elem:
-            for info in pkg:
-                if info.text in skip_versions:
+            name = pkg.attrib.get("name")
+            if not name:
+                continue
+            pkg_ns, _, pkg_name = name.rpartition("/")
+            purl = PackageURL(type="ebuild", name=pkg_name, namespace=pkg_ns)
+            safe_versions, affected_versions = GentooImporter.get_safe_and_affected_versions(pkg)
+
+            for version in safe_versions:
+                constraints.append(
+                    VersionConstraint(version=GentooVersion(version), comparator="=").invert()
+                )
+
+            for version in affected_versions:
+                constraints.append(
+                    VersionConstraint(version=GentooVersion(version), comparator="=")
+                )
+
+            if not constraints:
+                continue
+
+            yield AffectedPackage(
+                package=purl, affected_version_range=EbuildVersionRange(constraints=constraints)
+            )
+
+    @staticmethod
+    def get_safe_and_affected_versions(pkg):
+        # TODO : Revisit why we are skipping some versions in gentoo importer
+        skip_versions = {"1.3*", "7.3*", "7.4*"}
+        safe_versions = set()
+        affected_versions = set()
+        for info in pkg:
+            if info.text in skip_versions:
+                continue
+
+            if info.attrib.get("range"):
+                if len(info.attrib.get("range")) > 2:
                     continue
-                pkg_ns, pkg_name, = pkg.attrib[
-                    "name"
-                ].split("/")
-                purl = PackageURL(type="ebuild", name=pkg_name, version=info.text, namespace=pkg_ns)
 
-                if info.attrib.get("range"):
-                    if len(info.attrib.get("range")) > 2:
-                        continue
+            if info.tag == "unaffected":
+                # quick hack, to know whether this
+                # version lies in this range, 'e' stands for
+                # equal, which is paired with 'greater' or 'less'.
+                # All possible values of info.attrib['range'] =
+                # {'gt', 'lt', 'rle', 'rge', 'rgt', 'le', 'ge', 'eq'}, out of
+                # which ('rle', 'rge', 'rgt') are ignored, because they compare
+                # 'release' not the 'version'.
+                if "e" in info.attrib["range"]:
+                    safe_versions.add(info.text)
+                else:
+                    affected_versions.add(info.text)
 
-                if info.tag == "unaffected":
-                    # quick hack, to know whether this
-                    # version lies in this range, 'e' stands for
-                    # equal, which is paired with 'greater' or 'less'.
-                    # All possible values of info.attrib['range'] =
-                    # {'gt', 'lt', 'rle', 'rge', 'rgt', 'le', 'ge', 'eq'}, out of
-                    # which ('rle', 'rge', 'rgt') are ignored, because they compare
-                    # 'release' not the 'version'.
+            elif info.tag == "vulnerable":
+                if "e" in info.attrib["range"]:
+                    affected_versions.add(info.text)
+                else:
+                    safe_versions.add(info.text)
 
-                    if "e" in info.attrib["range"]:
-                        safe_purls.add(purl)
-                    else:
-                        affected_purls.add(purl)
-
-                elif info.tag == "vulnerable":
-                    if "e" in info.attrib["range"]:
-                        affected_purls.add(purl)
-                    else:
-                        safe_purls.add(purl)
-
-        return (affected_purls, safe_purls)
+        return safe_versions, affected_versions