+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Next release
 ----------------
 
 - We re-enabled support for the mozilla vulnerabilities advisories importer.
+- We re-enabled support for the gentoo vulnerabilities advisories importer.
 
 
 Version v31.1.1