+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Next release
 
 - We re-enabled support for the mozilla vulnerabilities advisories importer.
 - We re-enabled support for the gentoo vulnerabilities advisories importer.
+- We re-enabled support for the istio vulnerabilities advisories importer.
 
 
 Version v31.1.1